+++ Seasons.cs
@@ -20,7 +20,7 @@ namespace Seasons
     {
         public const string pluginID = "shudnal.Seasons";
         public const string pluginName = "Seasons";
-        public const string pluginVersion = "1.3.11";
+        public const string pluginVersion = "1.3.12";
 
         private readonly Harmony harmony = new Harmony(pluginID);
 
@@ -55,6 +55,7 @@ namespace Seasons
         public static ConfigEntry<TimerFormat> seasonsTimerFormatInRaven;
 
         public static ConfigEntry<bool> disableBloomInWinter;
+        public static ConfigEntry<Vector2> reduceSnowStormInWinter;
         public static ConfigEntry<bool> enableSeasonalItems;
         public static ConfigEntry<bool> preventDeathFromFreezing;
         public static ConfigEntry<bool> freezingSwimmingInWinter;
@@ -172,7 +173,7 @@ namespace Seasons
         private static HashSet<string> _GrassToControlSize = new HashSet<string>();
 
         private static int _instanceChangeIDShieldGeneratorCache;
-        private static readonly Dictionary<Vector3, bool> _cachedIgnoredPositions = new Dictionary<Vector3, bool>();
+        private static readonly Dictionary<Vector2, bool> _cachedIgnoredPositions = new Dictionary<Vector2, bool>();
 
         private static readonly Dictionary<string, GameObject> _treeRegrowthPrefabs = new Dictionary<string, GameObject>();
         public static int _treeRegrowthHaveGrowSpace = "Seasons_HaveGrowSpace".GetStableHashCode();
@@ -292,6 +293,10 @@ namespace Seasons
 
             disableBloomInWinter = config("Season", "Disable Bloom in Winter", defaultValue: true, "Force disables Bloom graphics setting while in Winter and restores it in other seasons (it will not change Graphics setting, only disables posteffect)." +
                                                                                                    "\nBloom in Winter is what makes you blind with that much of white.");
+            reduceSnowStormInWinter = config("Season", "Reduce SnowStorm particles in Winter", defaultValue: new Vector2(250, 1000), "Reduce SnowStorm particles emission rate and maximum amount. Vanilla values is 500:2000" +
+                                                                                                   "\nFirst parameter is emission rate and second is max particles amount." +
+                                                                                                   "\nHelps fps in Winter. Doesn't affect Mountains, Ashlands and DeepNorth." +
+                                                                                                   "\nSet to 0:0 to return Vanilla behaviour.");
             enableSeasonalItems = config("Season", "Enable seasonal items", defaultValue: true, "Enables seasonal (Halloween, Midsummer, Yule) items in the corresponding season");
             preventDeathFromFreezing = config("Season", "Prevent death from freezing", defaultValue: true, "Prevents death from freezing when not in mountains or deep north");
             seasonalStatsOutdoorsOnly = config("Season", "Seasonal stats works only outdoors", defaultValue: true, "Make seasonal stats works only outdoors");
@@ -312,6 +317,8 @@ namespace Seasons
             woodListToControlDrop.SettingChanged += (sender, args) => FillListsToControl();
             meatListToControlDrop.SettingChanged += (sender, args) => FillListsToControl();
             disableBloomInWinter.SettingChanged += (sender, args) => seasonState.UpdateWinterBloomEffect();
+            reduceSnowStormInWinter.SettingChanged += (sender, args) => ZoneSystemVariantController.SnowStormReduceParticlesChanged();
+            
 
             grassDefaultPatchSize = config("Season - Grass", "Default patch size", defaultValue: 10f, "Default size of grass patch (sparseness or how wide a single grass \"node\" is across the ground)" +
                                                                                                      "Increase to make grass more sparse and decrease to make grass more tight");
@@ -630,7 +637,8 @@ namespace Seasons
             if (WorldGenerator.instance == null)
                 return true;
 
-            if (_cachedIgnoredPositions.TryGetValue(position, out bool ignored))
+            Vector2 pos = new Vector2(position.x, position.z);
+            if (_cachedIgnoredPositions.TryGetValue(pos, out bool ignored))
                 return ignored;
 
             if (_cachedIgnoredPositions.Count > 15000)
@@ -640,7 +648,7 @@ namespace Seasons
                       WorldGenerator.IsDeepnorth(position.x, position.z) || 
                       WorldGenerator.instance.GetBaseHeight(position.x, position.z, menuTerrain: false) > WorldGenerator.mountainBaseHeightMin + 0.05f;
             
-            _cachedIgnoredPositions[position] = ignored;
+            _cachedIgnoredPositions[pos] = ignored;
             return ignored;
         }